requirement.txt
python-telegram-bot
requests
pandas
beautifulsoup4
tabulate
lxml
